Create draft2
working on logic for SW1 to scroll through reading type because I it's a little buggy, but for now this works! We definitely need a shorter delay function and it needs refining. I have PortE4 as the shuffle button and after you push it, it prints a card name to line 1 and most of the meaning on line 2.